Extract components common to team and user UBP
